PIlantir/backend/requirements.txt
paho-mqtt
python-dotenv
requests
bs4
requests
urllib3<2